Fix the frame-clock sample, time-scale/rate policy and recorder time base KI-51..KI-54 + KI-16; add the IFrameClock seam (WO-10)
Engine:
- core/IFrameClock.h: the frame time-source seam (IFrameClock + GlfwFrameClock); Application gains a
  test-only constructor overload taking the clock, Run()/RenderSingleFrame() sample through it. Inert in
  shipping: the default is glfwGetTime() sampled where it always was; no fake symbol in Cosmic.dll.
- KI-51: the clock is sampled in double and only the frame delta is narrowed to float (a float sample
  quantised every delta to the ulp of uptime: 0/2-tick jitter at 60 Hz, 7.8 ms at 24 h); m_AbsoluteTime
  accumulates in double behind the unchanged float getter (the float accumulator ran 3 % slow per day).
- KI-52 / KI-54 (policy): SetTimeScale accepts only finite values >= 0 - NaN, +/-inf and NEGATIVE are
  rejected with a warning and the previous scale kept (NaN poisoned the accumulator for good, +inf never
  left the drain loop, a negative scale was no rewind and left a hidden restart debt). The dead signed
  fixed-delta branch is removed: OnFixedUpdate always receives +1/Hz.
- KI-53: SetFixedTimestepHz rejects NaN (it survived std::clamp and stopped every fixed update until the
  next valid rate repaid the gap as one burst); every other value still clamps to [1, 1000].
- KI-16: DataRecorder::m_ElapsedTime accumulates in double (stored v1 timestamps stay float) - the float
  accumulator ended 16.9 s short after two nominal hours.
- Root CMakeLists: AnalysisSample (the X01 external-consumer specimen) on the project-scanner skip list
  in both modes; the skip-default migration's CACHE INTERNAL seed is guarded (INTERNAL implies FORCE, so
  an edited default could never migrate).

Contracts and docs: known-issues KI-51..KI-56 (+ KI-16 disposition), contracts.md s.5 record + new s.7
(global time scale and fixed-rate policy), numeric-bar-policy (WO-10 measurements: 2-h drift 0 steps,
24-h float quantisation -3 %), retained-feature-register M9/M11, docs/reference/core.md (IFrameClock,
SetTimeScale, SetFixedTimestepHz, GetAbsoluteTime) + manifest row, docs/guide/time-and-ticks.md.

+++ Cosmic/src/telemetry/DataRecorder.h
@@ -19,7 +19,7 @@
  * m_Records is populated exclusively during Register() on the main thread and
  * NEVER resized after that, so m_Records[id] dereferences are lock-free.
  * Each EntityRecord has its own mutex held only during Record() (< 1 µs).
- * m_ElapsedTime is std::atomic<float>; Tick() and RecordImpl() use relaxed
+ * m_ElapsedTime is std::atomic<double>; Tick() and RecordImpl() use relaxed
  * ordering — on x86 this is a plain mov with no barrier overhead.
  *
  * INTERNAL STORAGE — columnar layout
@@ -247,7 +247,10 @@ namespace Cosmic
         std::atomic<bool>       m_Flushing{ false };
         std::atomic<FlushState> m_FlushState{FlushState::Idle};
         std::function<void()> m_FlushWriteBarrier;
-        std::atomic<float>      m_ElapsedTime{ 0.0f }; // written by Tick (main), read by RecordImpl (workers)
+        // Accumulated in double (KI-16, WO-10): a float accumulator of 1/60 s ticks
+        // ended 16.9 s short after two hours. Still relaxed + lock-free on x64; the
+        // stored v1 timestamp is the correctly rounded float of the exact time.
+        std::atomic<double>     m_ElapsedTime{ 0.0 };  // written by Tick (main), read by RecordImpl (workers)
 
         // Autosave state (main-thread only — driven from Tick()).
         bool        m_AutosaveEnabled    = false;
